\name{DAIME-package}
\alias{DAIME-package}
\alias{DAIME}
\docType{package}
\title{
\packageTitle{DAIME}
}
\description{
\packageDescription{DAIME}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DAIME}
\packageIndices{DAIME}

}
\author{
\packageAuthor{DAIME}

Maintainer: \packageMaintainer{DAIME}
}

\references{
Hohmann, Niklas. 2018. Quantifying the Effects of Changing Deposition Rates and Hiatii on the Stratigraphic Distribution of Fossils. <doi:10.13140/RG.2.2.23372.51841>
}


\keyword{ package }
\keyword{ DAIME }
\keyword{ deposition rate }
\keyword{ stratigraphy }